Add '--policy' option to pdf_to_db command.

+++ api/ombpdf/management/commands/pdf_to_db.py
@@ -1,12 +1,12 @@
 from argparse import FileType
 from pathlib import Path
 
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandError
 
+from document.management.commands.import_xml_doc import fetch_policy
 from document.models import DocNode
 from ombpdf.document import OMBDocument
 from ombpdf.semdb import to_db
-from reqs.models import Policy
 
 
 class Command(BaseCommand):
@@ -14,12 +14,17 @@ class Command(BaseCommand):
 
     def add_arguments(self, parser):
         parser.add_argument('filename', type=FileType('rb'))
+        parser.add_argument('--policy',
+                            help='Policy id or number to associate')
 
     def handle(self, *args, **options):
         fp = options['filename']
-        omb_policy_id = Path(fp.name).stem.upper()
 
-        policy = Policy.objects.get(omb_policy_id=omb_policy_id)
+        policy_name = options['policy'] or Path(fp.name).stem.upper()
+        policy = fetch_policy(policy_name)
+
+        if policy is None:
+            raise CommandError(f"Policy '{policy_name}' not found.")
 
         self.stdout.write(f'Deleting any current document for "{policy}".')
         DocNode.objects.filter(policy=policy).delete()
